Added 10 directional, artificial maps for testing purposes in @biyonka's algorithm
These maps use the exponential kernel and are directional. Each map was generated with 50 points, and is bi-partite (if used with the masked kernel). To generate more maps just open the "nPartiteSpatialGenerator.nb", and run the whole thing. It will export csvs, and jpegs, into the data folder.

I still need to check that they work because it's been a long day, but they do seem to make sense.